# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Read Station API",
        "description": "API for Read Station",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/book-details": {
            "get": {
                "tags": [
                    "Admin / BookDetail"
                ],
                "summary": "Get all book details",
                "description": "Get all book details",
                "operationId": "getAllBookDetails",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Author id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all book details successfully!"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/book-details/{id}": {
            "get": {
                "tags": [
                    "Admin / BookDetail"
                ],
                "summary": "Get book detail",
                "description": "Get book detail",
                "operationId": "getBookDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of book detail",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get book detail successfully!"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Book detail not found!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/book-details/create": {
            "post": {
                "tags": [
                    "Admin / BookDetail"
                ],
                "summary": "Create book detail",
                "description": "Create book detail",
                "operationId": "createBookDetail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "sku_origin",
                                    "poster",
                                    "images",
                                    "book_version",
                                    "price",
                                    "hire_percent",
                                    "stock",
                                    "publish_date",
                                    "publishing_company_id",
                                    "issuing_company",
                                    "cardboard",
                                    "total_page",
                                    "language"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "string"
                                    },
                                    "sku_origin": {
                                        "type": "string"
                                    },
                                    "poster": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "book_version": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "hire_percent": {
                                        "type": "string"
                                    },
                                    "stock": {
                                        "type": "string"
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "publishing_company_id": {
                                        "type": "string"
                                    },
                                    "issuing_company": {
                                        "type": "string"
                                    },
                                    "cardboard": {
                                        "type": "string"
                                    },
                                    "total_page": {
                                        "type": "string"
                                    },
                                    "translator": {
                                        "type": "string"
                                    },
                                    "language": {
                                        "type": "string"
                                    },
                                    "book_size": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create book detail successfully!"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Create book detail failed!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/book-details/delete/{id}": {
            "delete": {
                "tags": [
                    "Admin / BookDetail"
                ],
                "summary": "Delete book detail",
                "description": "Delete book detail",
                "operationId": "deleteBookDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of book detail",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete book detail successfully!"
                    },
                    "400": {
                        "description": "Book detail is not found!"
                    },
                    "404": {
                        "description": "Book detail not found!"
                    },
                    "500": {
                        "description": "Delete book detail failed!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}